Fix memory issues.
ModuleFactory::create allocates memory. That must be deleted or kept in a smart pointer that will delete it.

module_wrapper_base.createModule() returns a unique_ptr. It was not stored in a local variable, so the memory that was allocated was immediately freed. The code still runs, accessing the freed memory, but this is undefined behavior.