Added module-dependency utility code in new utils directory.
Added utils directory to Makevars and Makevars.win and to the Doxygen
makefile and configuration files.

Modified the System constructor so that it re-orders steady-state
modules into a suitable evaluation order if possible and reports
whether this can or cannot be done.

Removed criterion 4 from the validate_system_inputs function.

Moved inclusion of the standard library set header file from
validate_system.cpp to validate_system.h so that there is no
inclusion-order dependency when validate_system.h is used.

Modified create_message function to prefix a mark to each message
which clearly indicates whether the message is reporting a validation
success or a validation failure.