Several changes to give Doxygen better, more useful defaults and give the documentation makefile more useful, regularly-named targets.
Also, a GitHub Workflow has been added to automatically build and
deploy the Doxygen documentation when a new commit is pushed to the
master branch.

Changes to Doxyfile include:

Changed the specification of the target Doxygen version in the lead
comment from 1.8.10 to 1.8.20.

Changed the project name to "The BioCro C++ Library" to clearly
indicate the scope of the Doxygen documentation.

Removed the PROJECT_NUMBER setting for now.

Set OUTPUT_DIRECTORY to doxygen_docs_complete.

Removed the OPTIMIZE_OUTPUT_FOR_C setting now that the project is in
C++.  Now, among other things, the documentation will refer to classes
as classes instead of as data structures.

Extract documentation of private members by default.

For clarity, stripped down FILE_PATTERNS setting to only what is
needed.  But added in patterns for Markdown files in case we want to
start using them.

Document the important subdirectories of src by default.

Don't include source code inline by default.

Added HTML_TIMESTAMP so it's easier for documentation users to tell
when the documentation was generated.

Set USE_MATHJAX to YES in case we want to start using it more
extensively.

Important changes to Makefile (renamed from makefile):

The Make file no longer uses a Doxygen configuration override file but
sets each overriden option on the command line using "echo".

A target to make only the module documentation has been added.

A few Doxygen options may now be specified on the Make command line.