Truncate long file paths
This is in response to the following output from R CMD check --as-cran:

Found the following non-portable file paths:
  BioCro/boost_1_71_0/boost/numeric/odeint/stepper/generation/generation_controlled_adams_bashforth_moulton.hpp
  BioCro/boost_1_71_0/boost/numeric/odeint/stepper/generation/generation_runge_kutta_cash_karp54_classic.hpp

Tarballs are only required to store paths of up to 100 bytes and cannot
store those of more than 256 bytes, with restrictions including to 100
bytes for the final component.
See section 'Package structure' in the 'Writing R Extensions' manual.