Stop hard-coding beta_PSII
This parameter in the C3 photosynthesis code had been hard-coded to 0.5. Several modules needed to be updated to include this new parameter in their inputs. All C3 crop model definitions needed to be updated to include the new parameter value, and the inputs for a few module tests also needed to be updated to include it.