Fixed small bug in the initialization of Dekker's method; + added continuity check
Changes
+ Dekker's method should set last = contrapoint, not last = best on initialization for the two point initialization.  Now, first iteration can be a secant update rather than always a bisection.
+ Added a continuity check to the convergence criteria in the bracketing methods and contrapoint methods.

For continuous functions, if bracketing method;s bracket shrinks to zero, then the method has indeed found a zero, but in practice, the tolerance for the `residual = 0`  does not always pass, creating spurious false convergences. However, non-continuous functions need to be disallowed. The solution here is to check to ensure we have always have three distinct points so that we can check whether the left and right secants agree (to within tolerance). If not, then it's a disconuity and otherwise its solution.